[SerialReader] Initial setup for stand-alone serial reader
This reader uses the "Strategy" design pattern.
This strategy must inherit from `SerialReaderStrategy` and implement this function:

```C++
virtual void checkForValidPacket() = 0;
```

Things to do in this function:

- Find start of packet
- check packet validity (e.g. CRC)
- call `markValidPacket`

The reader class exposes the following functions:

- `process()` To be called from the 10/sec or 50/sec loop
- `packetAvailable()` to check if a packet is available for processing.
- `readPacketByte` to read bytes from the packet buffer.
- `configureSoftwareSerial` To configure the reader to use a software serial port.
- `configureHardwareSerial` To configure to use one of the hardware serial ports. (setup of those ports must be done before calling this function)
- `write()` is just a pass-through to the active port.

N.B. This is not yet tested, not even compiled.
Just a start to make some kind of brain dump of this Idea I have about this reader.